+++ src/main/java/uos/uos25/user/dto/EmployeeInfo.java
@@ -1,4 +1,4 @@
-package uos.uos25.employee.dto;
+package uos.uos25.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
