+++ src/main/java/uos/uos25/user/dto/response/EmployeeHeadquarterGetResponseDTO.java
@@ -1,11 +1,11 @@
-package uos.uos25.employee.dto.response;
+package uos.uos25.user.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
-import uos.uos25.employee.entity.Employee;
+import uos.uos25.user.entity.Employee;
 
 @Data
-public class EmployeeGetResponseDTO {
+public class EmployeeHeadquarterGetResponseDTO {
     private Long employeeId;
     private String employeeName;
     private String position;
@@ -14,10 +14,9 @@ public class EmployeeGetResponseDTO {
     private String partTime;
     private String account;
     private String bank;
-    private Long shopId;
 
     @Builder
-    public EmployeeGetResponseDTO(
+    public EmployeeHeadquarterGetResponseDTO(
             Long employeeId,
             String employeeName,
             String position,
@@ -25,8 +24,7 @@ public class EmployeeGetResponseDTO {
             Integer salary,
             String partTime,
             String account,
-            String bank,
-            Long shopId) {
+            String bank) {
         this.employeeId = employeeId;
         this.employeeName = employeeName;
         this.position = position;
@@ -35,11 +33,10 @@ public class EmployeeGetResponseDTO {
         this.partTime = partTime;
         this.account = account;
         this.bank = bank;
-        this.shopId = shopId;
     }
 
-    public static EmployeeGetResponseDTO fromEntity(Employee employee) {
-        return EmployeeGetResponseDTO.builder()
+    public static EmployeeHeadquarterGetResponseDTO fromEntity(Employee employee) {
+        return EmployeeHeadquarterGetResponseDTO.builder()
                 .employeeId(employee.getEmployeeId())
                 .employeeName(employee.getEmployeeName())
                 .position(employee.getPosition())
@@ -48,7 +45,6 @@ public class EmployeeGetResponseDTO {
                 .partTime(employee.getPartTime().name())
                 .account(employee.getAccount())
                 .bank(employee.getBank())
-                .shopId(employee.getShop().getShopId())
                 .build();
     }
 }
