+++ src/main/java/uos/uos25/user/entity/EmployeeWorkingHistory.java
@@ -1,4 +1,4 @@
-package uos.uos25.employee.entity;
+package uos.uos25.user.entity;
 
 import java.time.LocalDateTime;
