+++ src/main/java/uos/uos25/user/exception/EmployeeNotFound.java
@@ -1,4 +1,4 @@
-package uos.uos25.employee.exception;
+package uos.uos25.user.exception;
 
 import uos.uos25.common.error.BusinessException;
 import uos.uos25.common.error.ErrorCode;
